--- platform_orchestrator/src/main/java/com/override/exception/LessonStructureException.java
@@ -1,6 +1,0 @@
-package com.override.exception;
-
-//Правильно понял логику, что BaseOrchestratorException для этого как раз и нужен?
-public class LessonStructureException extends BaseOrchestratorException{
-
-}
