+++ platform_orchestrator/src/main/resources/templates/lessons/core/1-introduction/1-step/1.html
